Refactor to ducks-modular-redux pattern for reducers/actions
