+++ src/ducks/web3.js
@@ -1,5 +1,5 @@
+// these will take in an action, have a default state set in the arguments and return a new state
 import {
-  INITIALIZE_GLOBAL_WEB3,
   SET_WEB3_CONNECTION_STATUS,
   SET_CURRENT_MASK_ADDRESS,
   METAMASK_LOCKED,
@@ -9,9 +9,11 @@ import {
   SET_NETWORK_MESSAGE,
   SET_BLOCK_TIMESTAMP,
   SET_EXCHANGE_TYPE,
+  INITIALIZE_GLOBAL_WEB3,
   TOGGLE_ABOUT,
   TOGGLE_INVEST
 } from '../constants';
+
 
 // this actions folder is actually full of action creators
 // your asynchronous calls are going to be in redux-thunk style action creators
@@ -106,3 +108,35 @@ export const initializeGlobalWeb3 = (web3) => ({
   type: INITIALIZE_GLOBAL_WEB3,
   web3
 });
+
+
+export default (state = {}, action) => {
+  const { connected, currentMaskAddress, metamaskLocked, interaction, factoryContract, networkMessage, timestamp, exchangeType, web3, aboutToggle, investToggle } = action
+  switch (action.type) {
+    case INITIALIZE_GLOBAL_WEB3:
+      return Object.assign({}, state, { web3: web3 });
+    case SET_WEB3_CONNECTION_STATUS:
+      return Object.assign({}, state, { connected: connected });
+    case SET_CURRENT_MASK_ADDRESS:
+      return Object.assign({}, state, { currentMaskAddress: currentMaskAddress });
+    case METAMASK_LOCKED:
+      return Object.assign({}, state, { metamaskLocked: metamaskLocked });
+    case METAMASK_UNLOCKED:
+      return Object.assign({}, state, { metamaskLocked: metamaskLocked });
+    case SET_INTERACTION_STATE:
+      return Object.assign({}, state, { interaction: interaction });
+    case FACTORY_CONTRACT_READY:
+      return Object.assign({}, state, { factoryContract: factoryContract});
+    case SET_NETWORK_MESSAGE:
+      return Object.assign({}, state, { networkMessage: networkMessage });
+    case SET_BLOCK_TIMESTAMP:
+      return Object.assign({}, state, { blockTimestamp: timestamp });
+    case SET_EXCHANGE_TYPE:
+      return Object.assign({}, state, { exchangeType: exchangeType });
+    case TOGGLE_ABOUT:
+      return Object.assign({}, state, { aboutToggle: aboutToggle })
+    case TOGGLE_INVEST:
+      return Object.assign({}, state, { investToggle: investToggle })
+    default: return state;
+  }
+}
